pyproject.toml
[project]
name = "movie_data_buiilder"
version = "0.1.0"
description = "continually writes and patches it together."
authors = [
    {name = "Tilman Holzhauer",email = "tilman.holzhauer@yahoo.de"}
]
license = {text = "Apache 2.0"}
readme = "README.md"
package-mode = false
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.77.0,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
